requirements.txt
flask==2.3.3
flask-login==0.6.2
werkzeug==2.3.7
firebase-admin==6.2.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
numpy==1.25.2
pandas==2.1.0
langchain==0.0.309
langchain-openai==0.0.2
openai==1.2.3
pypdf==3.16.1
docx2txt==0.8
tiktoken==0.5.1
spacy==3.6.1
gunicorn==21.2.0
scikit-learn==1.3.0
pydantic==2.4.2
chromadb==0.4.14
faiss-cpu==1.7.4